# connected-games-platform/docs/openapi.yaml
openapi: 3.0.3
info:
  title: PlayConnect API
  version: 2.0.0
  description: API REST didattica della piattaforma Giochi Connessi.
servers:
  - url: http://localhost:3000/api
components:
  securitySchemes:
    UserId:
      type: apiKey
      in: header
      name: X-User-Id
  schemas:
...
    GameType:
      type: object
      required: [name, description]
      properties:
        name: {type: string}
        description: {type: string}
        start_event: {type: string, default: MATCH_START}
        score_event_player1: {type: string, default: GOAL_PLAYER_1}
        score_event_player2: {type: string, default: GOAL_PLAYER_2}
        end_event: {type: string, default: MATCH_END}
        score_limit: {type: integer, nullable: true}
        supports_teams: {type: boolean}
...
    Tournament:
      type: object
      required: [name, game_type_id, participant_mode, locale_ids]
      properties:
        name: {type: string}
        game_type_id: {type: integer}
        participant_mode: {type: string, enum: [INDIVIDUAL, TEAM]}
        status: {type: string, enum: [DRAFT, ACTIVE, FINISHED]}
        locale_ids: {type: array, items: {type: integer}}
        team_ids: {type: array, items: {type: integer}}
        start_date: {type: string, format: date}
        end_date: {type: string, format: date}
security:
  - UserId: []
paths:
...
  /game-types:
    get:
      summary: Lista tipi di gioco
      responses: {'200': {description: Lista}}
    post:
      summary: Crea tipo di gioco
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/GameType'}
      responses: {'201': {description: Creato}}
...
  /matches/{id}/end:
    parameters: [{name: id, in: path, required: true, schema: {type: integer}}]
    post:
      summary: Termina partita
      responses: {'200': {description: Terminata}}
...
  /tournaments:
    get:
      summary: Lista tornei con locali e classifica
      responses: {'200': {description: Lista}}
    post:
      summary: Crea torneo multi-locale
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Tournament'}
      responses: {'201': {description: Creato}}
...